backend/requirements.txt
# Core Flask & WSGI
Flask==3.0.3
gunicorn==23.0.0
gevent==24.2.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
Flask-Migrate==4.0.7

# Authentication & Security
PyJWT==2.9.0
werkzeug==3.0.3
Flask-Talisman==1.1.0
cryptography==42.0.5

# CORS & Rate Limiting
flask-cors==4.0.1
Flask-Limiter==3.8.0

# Redis & Caching
redis==5.0.3
hiredis==2.3.2

# Celery for async tasks
celery==5.3.6
celery[redis]==5.3.6

# HTTP & DNS
requests==2.32.3
dnspython==2.6.1
urllib3==2.2.1
certifi==2024.2.2

# Security Scanning
cryptography==42.0.5
pyOpenSSL==24.1.0
sslyze==6.0.0

# Monitoring & Logging
sentry-sdk[flask]==1.45.0
prometheus-flask-exporter==0.23.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.1
validators==0.28.1

# Testing
pytest==8.1.1
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
locust==2.24.0
